+++ capital-bank-backend/src/main/java/com/springboot/capitalbank/repository/NoticeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface NoticeRepository  extends JpaRepository<Notice, Long> {
+public interface NoticeRepository  extends JpaRepository<Notice, Integer> {
 
     @Query(value = "from Notice n where CURDATE() BETWEEN noticBegDt AND noticEndDt")
     List<Notice> findAllActiveNotices();
